+++ src/main/java/com/pppp0722/vouchermanagement/entity/voucher/FixedAmountVoucher.java
@@ -9,10 +9,12 @@ public class FixedAmountVoucher implements Voucher {
     private final UUID voucherId;
     private final long amount;
     private final VoucherType type = VoucherType.FIXED_AMOUNT;
+    private final UUID memberId;
 
-    public FixedAmountVoucher(UUID voucherId, long amount) {
+    public FixedAmountVoucher(UUID voucherId, long amount, UUID memberId) {
         this.voucherId = voucherId;
         this.amount = amount;
+        this.memberId = memberId;
     }
 
     @Override
@@ -23,6 +25,11 @@ public class FixedAmountVoucher implements Voucher {
     @Override
     public long getAmount() {
         return amount;
+    }
+
+    @Override
+    public UUID getMemberId() {
+        return memberId;
     }
 
     @Override
