+++ src/main/java/com/pppp0722/vouchermanagement/entity/voucher/PercentDiscountVoucher.java
@@ -9,10 +9,12 @@ public class PercentDiscountVoucher implements Voucher {
     private final UUID voucherId;
     private final long amount;
     private final VoucherType voucherType = VoucherType.PERCENT_DISCOUNT;
+    private final UUID memberId;
 
-    public PercentDiscountVoucher(UUID voucherId, long amount) {
+    public PercentDiscountVoucher(UUID voucherId, long amount, UUID memberId) {
         this.voucherId = voucherId;
         this.amount = amount;
+        this.memberId = memberId;
     }
 
     @Override
@@ -23,6 +25,11 @@ public class PercentDiscountVoucher implements Voucher {
     @Override
     public long getAmount() {
         return amount;
+    }
+
+    @Override
+    public UUID getMemberId() {
+        return memberId;
     }
 
     @Override
