+++ src/main/java/com/pppp0722/vouchermanagement/entity/voucher/Voucher.java
@@ -10,4 +10,6 @@ public interface Voucher {
     VoucherType getType();
 
     long getAmount();
+
+    UUID getMemberId();
 }
