+++ src/main/java/com/pppp0722/vouchermanagement/io/Output.java
@@ -17,7 +17,9 @@ public interface Output {
 
     void printError(String error);
 
-    void printEntityTypeInputRequest();
+    void printEntityTypeInputRequest1();
+
+    void printEntityTypeInputRequest2();
 
     void printVoucherTypeInputRequest();
