+++ src/main/java/com/pppp0722/vouchermanagement/repository/member/MemberRepository.java
@@ -17,8 +17,4 @@ public interface MemberRepository {
     Member updateMember(Member member);
 
     Member deleteMember(Member member);
-
-    void updateWallet(UUID memberId, UUID walletId);
-
-    Optional<UUID> readWalletId(UUID memberId);
 }
