+++ src/main/java/com/pppp0722/vouchermanagement/repository/voucher/VoucherRepository.java
@@ -17,4 +17,6 @@ public interface VoucherRepository {
     Voucher updateVoucher(Voucher voucher);
 
     Voucher deleteVoucher(Voucher voucher);
+
+    List<Voucher> readVouchersByMemberId(UUID memberId);
 }
