+++ src/main/java/com/pppp0722/vouchermanagement/service/member/MemberService.java
@@ -8,7 +8,5 @@ public interface MemberService {
 
     Member createMember(UUID memberId, String name);
 
-    void updateWallet(UUID memberId, UUID walletId);
-
     List<Member> getAllMembers();
 }
