+++ src/main/java/com/pppp0722/vouchermanagement/service/member/MemberServiceImpl.java
@@ -27,11 +27,6 @@ public class MemberServiceImpl implements MemberService {
     }
 
     @Override
-    public void updateWallet(UUID memberId, UUID walletId) {
-        memberRepository.updateWallet(memberId, walletId);
-    }
-
-    @Override
     public List<Member> getAllMembers() {
         return memberRepository.readMembers();
     }
