+++ src/main/java/com/pppp0722/vouchermanagement/service/voucher/VoucherService.java
@@ -7,7 +7,9 @@ import java.util.UUID;
 
 public interface VoucherService {
 
-    void createVoucher(VoucherType voucherType, UUID voucherId, long amount);
+    void createVoucher(UUID voucherId, VoucherType voucherType, long amount, UUID memberId);
 
     List<Voucher> getAllVouchers();
+
+    List<Voucher> getVouchersByMemberId(UUID memberId);
 }
