+++ src/test/java/com/pppp0722/vouchermanagement/repository/member/JdbcMemberRepositoryTest.java
@@ -1,16 +1,16 @@
-package com.pppp0722.vouchermanagement.voucher.repository;
+package com.pppp0722.vouchermanagement.repository.member;
 
 import static com.wix.mysql.EmbeddedMysql.anEmbeddedMysql;
 import static com.wix.mysql.ScriptResolver.classPathScript;
 import static com.wix.mysql.config.Charset.UTF8;
 import static com.wix.mysql.config.MysqldConfig.aMysqldConfig;
 import static com.wix.mysql.distribution.Version.v5_7_latest;
-import static org.hamcrest.MatcherAssert.*;
-import static org.hamcrest.Matchers.*;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.samePropertyValuesAs;
 
-import com.pppp0722.vouchermanagement.entity.voucher.FixedAmountVoucher;
-import com.pppp0722.vouchermanagement.entity.voucher.Voucher;
-import com.pppp0722.vouchermanagement.repository.voucher.JdbcVoucherRepository;
+import com.pppp0722.vouchermanagement.entity.member.Member;
+
 import com.wix.mysql.EmbeddedMysql;
 import com.wix.mysql.config.MysqldConfig;
 import com.zaxxer.hikari.HikariDataSource;
@@ -42,10 +42,10 @@ import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 @SpringJUnitConfig
 @TestInstance(Lifecycle.PER_CLASS)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-@DisplayName("JdbcVoucherRepository 단위 테스트")
-class JdbcVoucherRepositoryTest {
+@DisplayName("JdbcMemberRepository 단위 테스트")
+class JdbcMemberRepositoryTest {
 
-    private static final Logger logger = LoggerFactory.getLogger(JdbcVoucherRepositoryTest.class);
+    private static final Logger logger = LoggerFactory.getLogger(JdbcMemberRepositoryTest.class);
 
     @Configuration
     @ComponentScan
@@ -75,11 +75,11 @@ class JdbcVoucherRepositoryTest {
     }
 
     @Autowired
-    JdbcVoucherRepository voucherRepository;
+    JdbcMemberRepository memberRepository;
 
     EmbeddedMysql embeddedMysql;
 
-    Voucher newVoucher;
+    Member newMember;
 
     @BeforeAll
     void setup() {
@@ -94,7 +94,7 @@ class JdbcVoucherRepositoryTest {
             .addSchema("test-voucher_mgmt", classPathScript("schema.sql"))
             .start();
 
-        newVoucher = new FixedAmountVoucher(UUID.randomUUID(), 10);
+        newMember = new Member(UUID.randomUUID(), "kim");
     }
 
     @AfterAll
@@ -104,35 +104,34 @@ class JdbcVoucherRepositoryTest {
 
     @Test
     @Order(1)
-    @DisplayName("Voucher 추가")
+    @DisplayName("Member 추가")
     public void testInsert() {
         try {
-            voucherRepository.createVoucher(newVoucher);
+            memberRepository.createMember(newMember);
         } catch (BadSqlGrammarException e) {
             logger.error("Got BadSqlGrammarException error code -> {}",
                 e.getSQLException().getErrorCode(), e);
         }
 
-        Optional<Voucher> retrievedVoucher = voucherRepository.readVoucher(
-            newVoucher.getVoucherId());
-        assertThat(retrievedVoucher.get(), samePropertyValuesAs(newVoucher));
+        Optional<Member> retrievedMember = memberRepository.readMember(
+            newMember.getMemberId());
+        assertThat(retrievedMember.get(), samePropertyValuesAs(newMember));
     }
 
     @Test
     @Order(2)
-    @DisplayName("Voucher 조회")
+    @DisplayName("Member 조회")
     public void testFindAll() {
-        List<Voucher> vouchers = voucherRepository.readVouchers();
-        assertThat(vouchers.isEmpty(), is(false));
+        List<Member> members = memberRepository.readMembers();
+        assertThat(members.isEmpty(), is(false));
     }
 
     @Test
     @Order(3)
-    @DisplayName("Voucher 삭제")
+    @DisplayName("Member 삭제")
     public void testDelete() {
-        Voucher deletedVoucher = voucherRepository.deleteVoucher(newVoucher);
-        Optional<Voucher> retrievedVoucher = voucherRepository.readVoucher(
-            deletedVoucher.getVoucherId());
-        assertThat(retrievedVoucher.isEmpty(), is(true));
+        Member deletedMember = memberRepository.deleteMember(newMember);
+        Optional<Member> retrievedMember = memberRepository.readMember(deletedMember.getMemberId());
+        assertThat(retrievedMember.isEmpty(), is(true));
     }
 }
