+++ src/test/java/com/pppp0722/vouchermanagement/repository/voucher/JdbcVoucherRepositoryTest.java
@@ -1,17 +1,15 @@
-package com.pppp0722.vouchermanagement.member.repository;
+package com.pppp0722.vouchermanagement.repository.voucher;
 
 import static com.wix.mysql.EmbeddedMysql.anEmbeddedMysql;
 import static com.wix.mysql.ScriptResolver.classPathScript;
 import static com.wix.mysql.config.Charset.UTF8;
 import static com.wix.mysql.config.MysqldConfig.aMysqldConfig;
 import static com.wix.mysql.distribution.Version.v5_7_latest;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.samePropertyValuesAs;
+import static org.hamcrest.MatcherAssert.*;
+import static org.hamcrest.Matchers.*;
 
-import com.pppp0722.vouchermanagement.entity.member.Member;
-
-import com.pppp0722.vouchermanagement.repository.member.JdbcMemberRepository;
+import com.pppp0722.vouchermanagement.entity.voucher.FixedAmountVoucher;
+import com.pppp0722.vouchermanagement.entity.voucher.Voucher;
 import com.wix.mysql.EmbeddedMysql;
 import com.wix.mysql.config.MysqldConfig;
 import com.zaxxer.hikari.HikariDataSource;
@@ -22,8 +20,12 @@ import javax.sql.DataSource;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
+import org.junit.jupiter.api.TestMethodOrder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,10 +36,15 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.jdbc.BadSqlGrammarException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
-class JdbcMemberRepositoryTest {
+@SpringJUnitConfig
+@TestInstance(Lifecycle.PER_CLASS)
+@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
+@DisplayName("JdbcVoucherRepository 단위 테스트")
+class JdbcVoucherRepositoryTest {
 
-    private static final Logger logger = LoggerFactory.getLogger(JdbcMemberRepositoryTest.class);
+    private static final Logger logger = LoggerFactory.getLogger(JdbcVoucherRepositoryTest.class);
 
     @Configuration
     @ComponentScan
@@ -67,11 +74,11 @@ class JdbcMemberRepositoryTest {
     }
 
     @Autowired
-    JdbcMemberRepository memberRepository;
+    JdbcVoucherRepository voucherRepository;
 
     EmbeddedMysql embeddedMysql;
 
-    Member newMember;
+    Voucher newVoucher;
 
     @BeforeAll
     void setup() {
@@ -86,7 +93,7 @@ class JdbcMemberRepositoryTest {
             .addSchema("test-voucher_mgmt", classPathScript("schema.sql"))
             .start();
 
-        newMember = new Member(UUID.randomUUID(), "kim");
+        newVoucher = new FixedAmountVoucher(UUID.randomUUID(), 10, UUID.randomUUID());
     }
 
     @AfterAll
@@ -96,34 +103,35 @@ class JdbcMemberRepositoryTest {
 
     @Test
     @Order(1)
-    @DisplayName("Member 추가")
+    @DisplayName("Voucher 추가")
     public void testInsert() {
         try {
-            memberRepository.createMember(newMember);
+            voucherRepository.createVoucher(newVoucher);
         } catch (BadSqlGrammarException e) {
             logger.error("Got BadSqlGrammarException error code -> {}",
                 e.getSQLException().getErrorCode(), e);
         }
 
-        Optional<Member> retrievedMember = memberRepository.readMember(
-            newMember.getMemberId());
-        assertThat(retrievedMember.get(), samePropertyValuesAs(newMember));
+        Optional<Voucher> retrievedVoucher = voucherRepository.readVoucher(
+            newVoucher.getVoucherId());
+        assertThat(retrievedVoucher.get(), samePropertyValuesAs(newVoucher));
     }
 
     @Test
     @Order(2)
-    @DisplayName("Member 조회")
+    @DisplayName("Voucher 조회")
     public void testFindAll() {
-        List<Member> members = memberRepository.readMembers();
-        assertThat(members.isEmpty(), is(false));
+        List<Voucher> vouchers = voucherRepository.readVouchers();
+        assertThat(vouchers.isEmpty(), is(false));
     }
 
     @Test
     @Order(3)
-    @DisplayName("Member 삭제")
+    @DisplayName("Voucher 삭제")
     public void testDelete() {
-        Member deletedMember = memberRepository.deleteMember(newMember);
-        Optional<Member> retrievedMember = memberRepository.readMember(deletedMember.getMemberId());
-        assertThat(retrievedMember.isEmpty(), is(true));
+        Voucher deletedVoucher = voucherRepository.deleteVoucher(newVoucher);
+        Optional<Voucher> retrievedVoucher = voucherRepository.readVoucher(
+            deletedVoucher.getVoucherId());
+        assertThat(retrievedVoucher.isEmpty(), is(true));
     }
 }
